moved STIR data containers stuff to stir_data_containers.*
