+++ src/xSTIR/cSTIR/CMakeLists.txt
@@ -20,7 +20,8 @@ set(CMAKE_POSITION_INDEPENDENT_CODE True)
 
 find_package(STIR REQUIRED)
 
-add_library(cstir cstir_p.cpp cstir_tw.cpp stir_x.cpp cstir.cpp)
+add_library(cstir 
+    cstir_p.cpp cstir_tw.cpp stir_data_containers.cpp stir_x.cpp cstir.cpp)
 target_include_directories(cstir PUBLIC
     "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>$<INSTALL_INTERFACE:include>")
